requirements.txt
fastapi
uvicorn
kafka-python
websockets